requirements.txt
pillow>=10.0
imagehash>=4.3
requests>=2.31
beautifulsoup4>=4.12
torch>=2.0
torchvision>=0.15
flask>=3.0
scikit-learn>=1.3
matplotlib>=3.7
tensorboard>=2.14
